requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose[cryptography]
passlib
jinja2
cryptography
websockets
pytest
pytest-asyncio
httpx
# Optional: For audio transcription (install separately if needed)
# faster-whisper